ref(shaolin-monks): implement block and improve combat callbacks

+++ shaolin-monks/update/gameplay/special_attacks.lua
@@ -1,11 +1,11 @@
 function handle_special_attack(p)
-  local handler, params = unpack_split(p.ca.handler, "#")
+  local vs, handler, params = get_vs(p), unpack_split(p.ca.handler, "#")
 
   if handler == "st_morph" then
     return st_morph(p, params)
   end
 
-  string_to_hash("fire_projectile,slide,bk_blade_fury,jc_high_green_bolt,jc_nut_cracker,jc_shadow_kick,jc_uppercut,jx_back_breaker,jx_gotcha,jx_ground_pound,kl_diving_kick,kl_hat_toss,kl_spin,kl_teleport,kn_fan_lift,kn_flying_punch,lk_bicycle_kick,lk_flying_kick,ml_ground_roll,ml_teleport_kick,rd_electric_grab,rd_teleport,rd_torpedo,rp_force_ball,rp_invisibility,sc_spear,sc_teleport_punch,sk_projectile,sk_sledgehammer,sz_freeze", { fire_projectile, slide, bk_blade_fury, jc_high_green_bolt, jc_nut_cracker, jc_shadow_kick, jc_uppercut, jx_back_breaker, jx_gotcha, jx_ground_pound, kl_diving_kick, kl_hat_toss, kl_spin, kl_teleport, kn_fan_lift, kn_flying_punch, lk_bicycle_kick, lk_flying_kick, ml_ground_roll, ml_teleport_kick, rd_electric_grab, rd_teleport, rd_torpedo, rp_force_ball, rp_invisibility, sc_spear, sc_teleport_punch, sk_projectile, sk_sledgehammer, sz_freeze })[handler](p)
+  string_to_hash("projectile,slide,bk_blade_fury,jc_high_green_bolt,jc_nut_cracker,jc_shadow_kick,jc_uppercut,jx_back_breaker,jx_gotcha,jx_ground_pound,kl_diving_kick,kl_hat_toss,kl_spin,kl_teleport,kn_fan_lift,kn_flying_punch,lk_bicycle_kick,lk_flying_kick,ml_ground_roll,ml_teleport_kick,rd_electric_grab,rd_teleport,rd_torpedo,rp_force_ball,rp_invisibility,sc_spear,sc_teleport_punch,sk_projectile,sk_sledgehammer,sz_freeze", { projectile, slide, bk_blade_fury, jc_high_green_bolt, jc_nut_cracker, jc_shadow_kick, jc_uppercut, jx_back_breaker, jx_gotcha, jx_ground_pound, kl_diving_kick, kl_hat_toss, kl_spin, kl_teleport, kn_fan_lift, kn_flying_punch, lk_bicycle_kick, lk_flying_kick, ml_ground_roll, ml_teleport_kick, rd_electric_grab, rd_teleport, rd_torpedo, rp_force_ball, rp_invisibility, sc_spear, sc_teleport_punch, sk_projectile, sk_sledgehammer, sz_freeze })[handler](p, vs)
 end
 
 function detect_special_attack(p, next_input)
@@ -32,9 +32,13 @@ function destroy_projectile(p)
   p.projectile = nil
 end
 
-function fire_projectile(p, max_t, before_callback, after_callback, collision_callback)
+function projectile(p)
+  create_projectile(p)
+end
+
+function create_projectile(p, max_t, before_callback, after_callback, collision_callback, reaction_callback)
   if not p.cap.has_fired_projectile then
-    p.projectile = p.projectile or string_to_hash("action,after_callback,before_callback,collision_callback,frames,max_t,params,sprites,x,y", { p.ca, after_callback, before_callback, collision_callback, 0, max_t, p.cap, p.character.projectile_sprites, p.x + sprite_w * p.facing, p.y + 5 - ceil(p.character.projectile_h / 2) })
+    p.projectile = p.projectile or string_to_hash("action,after_callback,before_callback,collision_callback,frames,max_t,params,reaction_callback,sprites,x,y", { p.ca, after_callback, before_callback, collision_callback, 0, max_t, p.cap, reaction_callback, p.character.projectile_sprites, p.x + sprite_w * p.facing, p.y + 5 - ceil(p.character.projectile_h / 2) })
     p.cap.has_fired_projectile = true
   end
 end
@@ -46,27 +50,39 @@ function update_projectile(p)
     p.projectile.t = (p.projectile.t or 0) + 1
 
     if p.projectile.before_callback then
-      p.projectile.before_callback(p)
+      p.projectile.before_callback(p, vs)
     end
 
     p.projectile.x += (p.projectile.x_speed or projectile_speed) * p.facing
     p.projectile.frames += 1
 
-    if not p.projectile.has_hit and has_collision(p.projectile.x, p.projectile.y, vs.x, vs.y, nil, 6) then
-      p.projectile.has_hit = true
-      deal_damage(action, params, vs)
-
-      if p.projectile.collision_callback then
-        p.projectile.collision_callback(p)
-      else
+    if not p.projectile.has_hit and not p.projectile.has_blocked and has_collision(p.projectile.x, p.projectile.y, vs.x, vs.y, nil, 6) then
+      if vs.ca == actions.block then
+        p.projectile.has_blocked = true
+        sfx(actions.block.hit_sfx)
+        deal_damage(vs, 1)
         destroy_projectile(p)
+      else
+        if p.projectile.action.hit_sfx then
+          sfx(p.projectile.action.hit_sfx)
+        end
+
+        p.projectile.has_hit = true
+        hit(action, params, vs)
+        vs.cap.reaction_callback = p.projectile.reaction_callback
+
+        if p.projectile.collision_callback then
+          p.projectile.collision_callback(p, vs)
+        else
+          destroy_projectile(p)
+        end
       end
     elseif is_limit_right(p.projectile.x) or is_limit_left(p.projectile.x) or (p.projectile.y > y_bottom_limit + sprite_h * 2) then
       destroy_projectile(p)
     end
 
     if p.projectile and p.projectile.after_callback then
-      p.projectile.after_callback(p)
+      p.projectile.after_callback(p, vs)
     end
 
     if p.projectile and p.projectile.max_t and p.projectile.t > p.projectile.max_t then
@@ -79,11 +95,13 @@ function update_projectile(p)
   end
 end
 
-function slide(p, ignore_break)
-  if p.t > 15 then
+function slide(p, _, max_t, ignore_break)
+  local max_t = max_t or 15
+
+  if p.t > max_t then
     finish_action(p)
   else
-    if p.t <= (ignore_break and 15 or 8) then
+    if p.t <= (ignore_break and max_t or max_t / 2) then
       move_x(p, offensive_speed)
     end
 
@@ -91,9 +109,7 @@ function slide(p, ignore_break)
   end
 end
 
-function teleport(p, next_action, next_action_params, teleport_callback)
-  local vs = get_vs(p)
-
+function teleport(p, vs, next_action, next_action_params, teleport_callback)
   if not p.cap.has_teleported then
     if p.y < y_bottom_limit + (sprite_h * 2) + stroke_width then
       move_y(p, jump_speed)
@@ -106,32 +122,30 @@ function teleport(p, next_action, next_action_params, teleport_callback)
   end
 end
 
-function bk_blade_fury(p)
-  if not p.cap.has_hit then
-    if p.t > 30 then
-      finish_action(p)
-    else
-      attack(
-        p, function(p)
-          local vs = get_vs(p)
-          move_x(vs, -1)
-          move_y(vs, -1)
-          vs.cap.reaction_callback = function(p)
-            if p.t > 30 then
-              finish_action(get_vs(p))
-              finish_action(p, actions.thrown_backward)
-            elseif p.t % 5 == 0 then
-              spill_blood(p)
-            end
-          end
+function bk_blade_fury(p, vs)
+  if p.t > 30 then
+    finish_action(p)
+  else
+    attack(
+      p,
+      function(p, vs)
+        move_x(vs, -1)
+        move_y(vs, -1)
+      end,
+      function(p, vs)
+        if p.t > 30 then
+          finish_action(vs)
+          finish_action(p, actions.thrown_backward)
+        elseif p.t % 5 == 0 then
+          spill_blood(p)
         end
-      )
-    end
+      end
+    )
   end
 end
 
 function jc_high_green_bolt(p)
-  fire_projectile(
+  create_projectile(
     p, nil, function(p)
       if p.projectile.top_height_reached then
         p.projectile.y *= 1.05
@@ -162,27 +176,26 @@ function jc_uppercut(p)
     end
   else
     attack(p)
-    move_x(p, 1)
+    move_x(p, offensive_speed)
     move_y(p, -offensive_speed)
     p.cap.top_height_reached = p.y <= y_upper_limit
   end
 end
 
-function jx_back_breaker(p)
-  attack(p)
+function jx_back_breaker(p, vs)
+  attack(
+    p, function(p, vs)
+      vs.y = p.y - (vs.caf > 8 and 5 or 6)
+      vs.x = p.x
+    end
+  )
 
-  if p.cap.has_hit then
-    local vs = get_vs(p)
-    vs.y = p.y - (vs.caf > 8 and 5 or 6)
-    vs.x = p.x
-  else
+  if not p.cap.has_hit then
     setup_next_action(p, "jump", p.cap, true)
   end
 end
 
-function jx_gotcha(p)
-  local vs, max_punches_landed = get_vs(p)
-
+function jx_gotcha(p, vs)
   if p.cap.has_hit then
     p.cap.punches, p.cap.max_punches = p.cap.punches or 1, p.cap.max_punches or 2
     local is_last_punch = p.cap.punches >= p.cap.max_punches
@@ -211,22 +224,21 @@ function jx_gotcha(p)
   end
 end
 
-function jx_ground_pound(p)
+function jx_ground_pound(p, vs)
   if p.t > 20 then
     finish_action(p)
   else
     attack(
-      p, function(p)
-        local vs = get_vs(p)
-        vs.cap.reaction_callback = function(p)
-          move_x(p, -1)
-          if p.t > 15 then
-            finish_action(p)
-          end
+      p,
+      nil,
+      function(p)
+        move_x(p, -1)
+        if p.t > 15 then
+          finish_action(p)
         end
       end,
-      function(p)
-        return get_vs(p).y == y_bottom_limit
+      function(p, vs)
+        return vs.y == y_bottom_limit
       end
     )
   end
@@ -237,7 +249,7 @@ function kl_diving_kick(p)
 end
 
 function kl_hat_toss(p)
-  fire_projectile(
+  create_projectile(
     p, nil, function(p)
       if btn(⬆️, p.id) then
         p.projectile.y -= 0.75
@@ -265,9 +277,13 @@ function kl_spin(p)
   end
 end
 
-function kl_teleport(p)
+function kl_teleport(p, vs)
   teleport(
-    p, actions.jump, nil, function(p, vs)
+    p,
+    vs,
+    actions.jump,
+    nil,
+    function(p, vs)
       p.x = vs.x - sprite_w * vs.facing
       p.facing *= -1
       fix_players_orientation()
@@ -276,18 +292,20 @@ function kl_teleport(p)
 end
 
 function kn_fan_lift(p)
-  fire_projectile(
-    p, 30, function(p)
+  create_projectile(
+    p,
+    30,
+    function(p)
       p.projectile.x_speed = 0
-    end, nil, function(p)
-      local vs = get_vs(p)
-      vs.cap.reaction_callback = function(p)
-        if p.t > 60 then
-          finish_action(p, actions.fall)
-        elseif p.t < 10 then
-          move_x(vs, -1)
-          move_y(vs, -1)
-        end
+    end,
+    nil,
+    function() end,
+    function(p)
+      if p.t > 60 then
+        finish_action(p, actions.fall)
+      elseif p.t < 10 then
+        move_x(p, -1)
+        move_y(p, -1)
       end
     end
   )
@@ -313,10 +331,8 @@ function kn_flying_punch(p)
   end
 end
 
-function lk_bicycle_kick(p)
+function lk_bicycle_kick(p, vs)
   if p.cap.has_hit then
-    local vs = get_vs(p)
-
     if vs.t > 30 then
       finish_action(p)
       finish_action(vs)
@@ -336,52 +352,43 @@ function lk_bicycle_kick(p)
 end
 
 function lk_flying_kick(p)
-  if p.t > 20 then
-    finish_action(p)
-  else
-    move_x(p, offensive_speed)
-    attack(p, finish_action)
-  end
+  slide(p, nil, 20)
 end
 
 function ml_ground_roll(p)
-  slide(p, true)
+  slide(p, nil, 20, true)
 end
 
-function ml_teleport_kick(p)
+function ml_teleport_kick(p, vs)
   teleport(
-    p, actions.jump_kick, { is_landing = true }, function(p, vs)
-      p.x = get_vs(p).x - (sprite_w / 2)
+    p, vs, actions.jump_kick, { is_landing = true }, function(p, vs)
+      p.x = vs.x - (sprite_w / 2)
       p.y = y_upper_limit
     end
   )
 end
 
-function rd_electric_grab(p)
-  local vs = get_vs(p)
-
-  if p.cap.has_hit then
-    if vs.t > 30 then
-      finish_action(p)
-      finish_action(vs, actions.thrown_backward)
-    end
+function rd_electric_grab(p, vs)
+  if not p.cap.has_hit and p.t > 15 then
+    finish_action(p)
   else
-    if p.t > 15 then
-      finish_action(p)
-    else
-      attack(
-        p, function(p, vs)
-          move_x(vs, -1)
-          move_y(vs, -1)
+    attack(
+      p, nil, function(p, vs)
+        if p.t > 30 then
+          finish_action(vs)
+          finish_action(p, actions.fall)
+        elseif p.t < 2 then
+          move_x(p, -1)
+          move_y(p, -1)
         end
-      )
-    end
+      end
+    )
   end
 end
 
-function rd_teleport(p)
+function rd_teleport(p, vs)
   teleport(
-    p, actions.idle, nil, function(p, vs)
+    p, vs, actions.idle, nil, function(p, vs)
       p.x = vs.x - sprite_w * vs.facing
       p.y = vs.y
       p.facing *= -1
@@ -390,25 +397,26 @@ function rd_teleport(p)
   )
 end
 
-function rd_torpedo(p)
-  local vs = get_vs(p)
-
-  if p.cap.has_hit then
-    if is_limit_left(vs.x) or is_limit_right(vs.x) then
-      finish_action(p, actions.jump, { direction = backward })
-      finish_action(vs, actions.prone, nil)
-    else
-      move_x(p, offensive_speed)
-      move_x(vs, -offensive_speed)
-    end
-  else
+function rd_torpedo(p, vs)
+  if p.t < 30 then
     move_x(p, offensive_speed)
-    attack(p)
+    attack(
+      p, nil, function(p, vs)
+        if is_limit_left(p.x) or is_limit_right(p.x) then
+          finish_action(vs, actions.jump, { direction = backward })
+          finish_action(p, actions.prone, nil)
+        else
+          move_x(p, -offensive_speed)
+        end
+      end
+    )
+  else
+    finish_action(p)
   end
 end
 
 function rp_force_ball(p)
-  fire_projectile(
+  create_projectile(
     p, 90, function(p)
       p.projectile.x_speed = 0.5
     end
@@ -430,15 +438,14 @@ function st_morph(p, id)
 end
 
 function sc_spear(p)
-  fire_projectile(
+  create_projectile(
     p,
     nil,
     function(p)
       p.projectile.has_rope = true
     end,
-    function(p)
+    function(p, vs)
       if p.projectile.has_hit then
-        local vs = get_vs(p)
         if vs.t > 10 then
           if has_collision(p.x, p.y, vs.x, vs.y) then
             finish_action(p)
@@ -463,8 +470,6 @@ function sc_spear(p)
 end
 
 function sc_teleport_punch(p)
-  local vs = get_vs(p)
-
   if not p.cap.has_teleported then
     if not p.cap.has_changed_orientation then
       p.facing *= -1
@@ -484,7 +489,7 @@ function sc_teleport_punch(p)
 end
 
 function sk_projectile(p)
-  fire_projectile(
+  create_projectile(
     p, nil, function(p)
       p.projectile.x_speed = projectile_speed * 1.5
     end
@@ -492,7 +497,7 @@ function sk_projectile(p)
 end
 
 function sk_sledgehammer(p)
-  fire_projectile(
+  create_projectile(
     p, nil, function(p)
       p.projectile.x_speed = 0
     end,
@@ -529,9 +534,8 @@ function sk_sledgehammer(p)
 end
 
 function sz_freeze(p)
-  fire_projectile(
-    p, nil, nil, nil, function(p)
-      local vs = get_vs(p)
+  create_projectile(
+    p, nil, nil, nil, function(p, vs)
       if not is_st_eq(vs, "frozen") then
         vs.st_timers.frozen = 60
       else
